code/smoke_detector/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
idna==2.10
imageio==2.9.0
imutils==0.5.4
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector-python==8.0.25
networkx==2.5.1
numpy==1.19.5
Pillow==8.2.0
protobuf==3.17.0
pykwalify==1.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.20.1
scipy==1.5.4
SimpleITK==2.0.2
six==1.16.0
threadpoolctl==2.1.0
tifffile==2020.9.3
urllib3==1.26.4
watchdog==2.1.1
